// src/app/shared/enums/allergen.enum.ts
export enum Allergen {
  
  GLUTEN = 'GLUTEN',
  CRUSTACEANS = 'CRUSTACEANS',
  EGGS = 'EGGS',
  FISH = 'FISH',
  PEANUTS = 'PEANUTS',
  SOYBEANS = 'SOYBEANS',
  MILK = 'MILK',
  NUTS = 'NUTS',
  CELERY = 'CELERY',
  MUSTARD = 'MUSTARD',
  SESAME = 'SESAME',
  SULFITES = 'SULFITES',
  LUPIN = 'LUPIN',
  MOLLUSCS = 'MOLLUSCS'
}